Split editor extensions into multiple module loaded files. Re-using editor/deploy.ts for cmds. Fixing field ports
